Improve image processing and comparison in Twitch drops detection
In this commit, numerous changes have been made to the ImageProcessor.cs and ImageComparator.cs to refine the process of detecting Twitch drops in the 'Rust' game. A comprehensive color filter has been added to the ImageProcessor, which should enhance the image matching accuracy by ignoring certain colors that are prevalent in the game screen's UI. Additional methods to reduce noise and better recognize images within a grid have also been introduced.

In the ImageComparator, we have now incorporated filtering into the comparison process, updated thresholds in comparison methods like ORB and SSIM, and introduced new comparison algorithms - MatchTemplate and AbsDiff - to improve the overall comparison efficiency and the proportion of successful matches.

The changes aim to make the app more robust and reliable, leading to a better user experience.